"Complete"-ed NoDL conversion API
Concerns:
* Enclave paths are not well specified. Is it fair to make an
assumption and default to `"[node_name][executable_name]"`?
* Node namespaces are also not well specified. It certainly is
not fair to assume all nodes belong to the unspecified `"/"`
general namespace.

Both of the above concerns can be addressed by expanding the
definition of the node interface description language. This
discussion probably has a place in ros2/design#266.

TO-DO:
* Verify whether default enclave paths should be left blank
or should be set to an unqualified general path ("/")
* Writing unit and functional tests for the `nodl_to_policy`
utility.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>