Remove use of leading underscore for submodules
The leading underscore naming convention is used for "private" symbols
in Python, and as such, was being used inappropriately (and excessively)
in this package. This commit addresses that concern and only keeps true
implementation details private.

A sid-effect of this PR is that automated document generation tools such
as `sphinx-apidoc` can generate more detailed documentation for the
package.

Signed-off-by: Abrar Rahman Protyasha <abrar@openrobotics.org>